Some fixes
Some fixes to the code made with the clang --analyze flag

The operating system doesn't build with clang anymore but can still
diagnose some possible issues through static analysis.

+++ src/memory.c
@@ -108,7 +108,7 @@ void memory_init(void *start, u32 bytes) {
 	memory_size = bytes;
 	const u32 heap_size = 128 * 1024;
 	
-	if (alloc_ptr < start)
+	if (alloc_ptr < (u32)start)
 		alloc_ptr = (u32)start;
 	current_directory = kmalloc_a(sizeof(struct PageDirectory));
 	memset(current_directory, 0, sizeof(struct PageDirectory));
